third_party/WebKit/LayoutTests/fast/dom/location-missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldThrow("document.location.assign()");
shouldThrow("document.location.replace()");
</script>
</body>
</html>
